*: move test-only code into test files
Move some code that's only used within test builds into test files. In
the case of `(*memTable).close`, remove it altogether because it's a
no-op.

For some reason I don't understand, the `staticcheck` linter surfaced
these on a work-in-progress branch of mine that didn't touch anything
related to these.

```
mem_table.go:154:20: func (*memTable).get is unused (U1000)
mem_table.go:265:20: func (*memTable).close is unused (U1000)
mem_table.go:270:20: func (*memTable).empty is unused (U1000)
read_compaction_queue.go:29:32: func (*readCompactionQueue).at is unused (U1000)
sstable/compression_cgo.go:24:7: const useStandardZstdLib is unused (U1000)
sstable/reader.go:2060:18: func (*Reader).get is unused (U1000)
```
