+++ sstable/reader_test.go
@@ -30,6 +30,55 @@ import (
 	"golang.org/x/exp/rand"
 )
 
+// get is a testing helper that simulates a read and helps verify bloom filters
+// until they are available through iterators.
+func (r *Reader) get(key []byte) (value []byte, err error) {
+	if r.err != nil {
+		return nil, r.err
+	}
+
+	if r.tableFilter != nil {
+		dataH, err := r.readFilter()
+		if err != nil {
+			return nil, err
+		}
+		var lookupKey []byte
+		if r.Split != nil {
+			lookupKey = key[:r.Split(key)]
+		} else {
+			lookupKey = key
+		}
+		mayContain := r.tableFilter.mayContain(dataH.Get(), lookupKey)
+		dataH.Release()
+		if !mayContain {
+			return nil, base.ErrNotFound
+		}
+	}
+
+	i, err := r.NewIter(nil /* lower */, nil /* upper */)
+	if err != nil {
+		return nil, err
+	}
+	ikey, value := i.SeekGE(key)
+
+	if ikey == nil || r.Compare(key, ikey.UserKey) != 0 {
+		err := i.Close()
+		if err == nil {
+			err = base.ErrNotFound
+		}
+		return nil, err
+	}
+
+	// The value will be "freed" when the iterator is closed, so make a copy
+	// which will outlast the lifetime of the iterator.
+	newValue := make([]byte, len(value))
+	copy(newValue, value)
+	if err := i.Close(); err != nil {
+		return nil, err
+	}
+	return newValue, nil
+}
+
 // iterAdapter adapts the new Iterator API which returns the key and value from
 // positioning methods (Seek*, First, Last, Next, Prev) to the old API which
 // returned a boolean corresponding to Valid. Only used by test code.
@@ -108,18 +157,18 @@ func (i *iterAdapter) SetBounds(lower, upper []byte) {
 func TestReader(t *testing.T) {
 	writerOpts := map[string]WriterOptions{
 		// No bloom filters.
-		"default": WriterOptions{},
-		"bloom10bit": WriterOptions{
+		"default": {},
+		"bloom10bit": {
 			// The standard policy.
 			FilterPolicy: bloom.FilterPolicy(10),
 			FilterType:   base.TableFilter,
 		},
-		"bloom1bit": WriterOptions{
+		"bloom1bit": {
 			// A policy with many false positives.
 			FilterPolicy: bloom.FilterPolicy(1),
 			FilterType:   base.TableFilter,
 		},
-		"bloom100bit": WriterOptions{
+		"bloom100bit": {
 			// A policy unlikely to have false positives.
 			FilterPolicy: bloom.FilterPolicy(100),
 			FilterType:   base.TableFilter,
